rkwarddev: began to replace the static i18n stuff with the method to be introduced with RKWard 0.6.3. this breaks compatibility with rkwarddev 0.06-4, but it wasn't ever really used except for a proof of concept anyway.
this commit brings the "i18n" argument to (hopefully) every function where it could be used. it also accepts a simple character string, which will be taken as context info.

there's a new function rk.i18n.comment() and a rewrite of i18n(), which is now a convenient all-in-one wrapper for the JS functions i18n(), i18nc(), i18np() and i18ncp().