updated install logic for Mac (and a select general changes)
This implements a first draft of a minimal build and install
procedure on Mac, meaning with as little changes as possible from
a standard Linux install but still providing a complete "Mac
experience".
A hook is provided for a future optional build as a standalone,
all-encompassing .app bundle
The application icon is generated on the fly from the available
png and svgz images (with a new enough ECM library).