Rename rinterface.h in order to avoid include file clash.
Sigh. clang considers Mac to be case-insensitive, and so will consider
rinterface.h the same as Rinterface.h.

Further clang and cmake (don't know which is to blame) think it's fun to look
for <include.h> in the local dirs, first, then in the system path. Thus,

would include our rinterface.h, instead of the R header.