Remove the explicit intialization of a compiler, as it breaks with R 3.5.0 (https://bugs.debian.org/cgi-bin/bugreport.cgi?bug=900591).
This logic was added in 11/2007 / RKward 0.4.8a, when R was specifically built
using GCC 4.2 on debian but not Ubuntu, while the default compiler on both was still GCC 4.1.

It was meant to ensure a successful build on both systems.

Not sure whether a similar solution may be needed again, in the future, but we'll worry about that,
if and when it happens.