Fix message extraction with scripty
subprocess.Popen and bash -c are now used because the default value
of XGETTEXT set by scripty is an function defined in the environment.
Using subprocess.call with shell=True uses sh, which does not pass
the function defined with export -f, so let's be explicit
and depend on bash.