import_translations.py: remove comment stripping
Removing the comments from the po files is not exactly a good idea,
considering that also the header (containing e.g. copyright and
licenses) is stripped. In general, the risk of editing translations in
the wrong place (e.g. from a release tarball) has never been a big
concern, at least in the KDE community.