Rework in-source layout of plugin translations
Instead of copying all the plugin catalogs directly under 'i18n/po'
with the language as part of the filename, copy them under a
subdirectory for their language. This has few advantages:
- there is no need to rename them when copying in the sources, and when
  compiling them
- each language gets its own CMakeLists.txt (copied from a static
  template), so it is possible to glob for .po files, and use the
  standard GETTEXT_PROCESS_PO_FILES() macro of the Gettext cmake module
- a lot of the current manual code to build translations, including
  renaming, custom targets, etc is removed
- each .po file is always compiled, so there is no need to execute
  msgfmt first to get its statistics (thus the build is faster)

The only behaviour change is that now all the shipped .po files are
compiled and installed, no matter their completion status.
The previous code used to not compile a .po file with less than 80% of
translated messages; this is something that has not been generally done
for many years in KDE, as the completion percentage does not say much
about the actual usefulness of a translation. If something like this
is needed again, then a better way is to simply not copy translations
in the sources, rather than ignoring them.