update_plugin_messages.py: tweak temporary cpp files
One of the tools in scripty is repack-pot.pl, which replaces the
references to temporary sources used to collect strings from non-source
files (like XML, UI, etc) with the actual references, using the i18n
comments in the temporary sources. The only name of temporary source
recognized is "rc.cpp", as any other name can potentially be a real
source. The temporary sources created are "$po_id.pot.cpp", and thus
trigger warnings of repack-pot.pl.

As a solution/workaround, rename these "$po_id.pot.cpp" to "rc.cpp"
just before invoking xgettext, which now can track the location of the
strings. The end result is the very same.