Remove old (custom) i18n installation handling.
Note: Compiled po files for plugins shipped with the main distribution are now installed
in a regular system path. RKWard will still try to find them relative to the .pluginmap files
(as is needed for external plugins), but as the system path is always used as a fallback, no
adjustments to this mechanism are needed.