added code to make the removal of NULL default values not a breaking change to affected functions
it could be that some other code or scripts use the old NULL values explicitly as a feature.
now, if the arguments in questions are given a NULL value, a warning will ne shown that
this is now deprecated, and the verable will be made missing so the rest of the function
behaves as expected.